+++ scripts/startup-dev.sh
@@ -6,17 +6,17 @@
 
 echo "EXECUTING START-UP SCRIPT IN DEVELOPMENT"
 
-chmod +rw pnpm-lock.yaml \
-    .babelrc.js \
-    .eslintignore \
-    .eslintrc \
-    .prettierignore \
-    .prettierrc.json \
-    commitlint.config.js \
-    lint-staged.config.js \
-    tsconfig.json \
-    tsoa.json
+# chmod +rw pnpm-lock.yaml \
+#     .babelrc.js \
+#     .eslintignore \
+#     .eslintrc \
+#     .prettierignore \
+#     .prettierrc.json \
+#     commitlint.config.js \
+#     lint-staged.config.js \
+#     tsconfig.json \
+#     tsoa.json
 
-pnpm i
+# pnpm i
 # pnpm dev:nodemon
-pnpm start:ts
+pnpm dev:server