+++ src/modules/cli/account.ts
@@ -140,7 +140,7 @@ export async function cliAuthenticate(options: InputOptions) {
 		const user = await cliLogin(options);
 
 		if (!user) {
-			logError(`Can't login to the build server...`);
+			logError(`Failed to login: User not found.`);
 			return;
 		}
 
@@ -169,7 +169,7 @@ export async function cliAuthenticate(options: InputOptions) {
 	// log(`user :>>`, user);
 
 	if (!status || isEmpty(user)) {
-		logError(`Authentication failed.`, messages);
+		// logError(`Authentication failed.`, messages);
 		if (buildServerUrl) user = await continueToLoginStep(buildServerUrl);
 	}